I'm changing the structure. Now everything is in one file: cryptopals.py
I'm keeping around the old c0?_*.py files for now, because at least the
one for challenge 1 has some potentially useful code in it. I'd like to
convert that code to some sort of note in notes.markdown before removing
it. For example, it has my base64 encoder that doesn't rely on the bas64
python module.