OK, I think I have found a way forward.
I'm using letter frequency to score candidates.

I found a chart on wikipedia that counts a-z, and it's working sorta,
but it's not perfect. It gets more than half of the tchunks in my
example correct, but it's missing a lot.

At least a few of them would be fixed by adding a space to the
consideration... I'm going to work on that next. I have to figure out
an approximate value for spaces to go into
SingleCharcandidate.avghistogram though, because I can't find any data
that already exists on it.